crud de companias

+++ src/main/java/com/jaxws/controllers/CompanyController.java
@@ -1,6 +1,7 @@
 package com.jaxws.controllers;
 
 import com.jaxws.dtos.CompanyDto;
+import com.jaxws.dtos.Request;
 import com.jaxws.dtos.Response;
 import com.jaxws.services.CompanyService;
 import jakarta.jws.HandlerChain;
@@ -12,10 +13,10 @@ import jakarta.xml.bind.annotation.XmlRootElement;
 
 @WebService(name = "CompanyController", serviceName = "CompanyController")
 @XmlRootElement(name = "COMPANY_RESPONSE")
-@HandlerChain(file = "handler-chain.xml")
 public class CompanyController {
 
     private final CompanyService companyService;
+    private static final String ERROR_MESSAGE = "Missing or invalid Authorization header";
 
     public CompanyController() {
         this.companyService = new CompanyService();
@@ -24,23 +25,47 @@ public class CompanyController {
 
     @WebMethod(operationName = "GET_COMPANIES", action = "GET_COMPANIES")
     @XmlElement(name = "COMPANY")
-    public Response getCompanies() {
-        return companyService.getCompanies();
+    public Response getCompanies(Request request) {
+
+        if (!request.getAuthToken().isEmpty() && request.validateToken()) {
+            return companyService.getCompanies();
+
+        } else {
+            throw new SecurityException(ERROR_MESSAGE);
+        }
+
     }
 
     @WebMethod(operationName = "CREATE_COMPANY", action = "CREATE_COMPANY")
     public Response registerCompany(@WebParam(name = "COMPANY") CompanyDto company) {
-        return companyService.createCompany(company);
+        if (!company.getAuthToken().isEmpty() && company.validateToken()) {
+            return companyService.createCompany(company);
+
+        } else {
+            throw new SecurityException(ERROR_MESSAGE);
+        }
     }
 
     @WebMethod(operationName = "UPDATE_COMPANY", action = "UPDATE_COMPANY")
-    public Response updateCompany(@WebParam(name = "COMPANY") CompanyDto company, @WebParam(name = "ID") int id)  {
-        return companyService.updateCompany(company, id);
+    public Response updateCompany(@WebParam(name = "COMPANY") CompanyDto company, @WebParam(name = "ID") int id) {
+
+        if (!company.getAuthToken().isEmpty() && company.validateToken()) {
+            return companyService.updateCompany(company, id);
+        } else {
+            throw new SecurityException(ERROR_MESSAGE);
+        }
+
     }
 
     @WebMethod(operationName = "DELETE_COMPANY", action = "DELETE_COMPANY")
-    public Response deleteCompany(@WebParam(name = "ID") int id)  {
-        return companyService.deleteCompany(id);
+    public Response deleteCompany(@WebParam(name = "ID") int id, Request request) {
+
+        if (!request.getAuthToken().isEmpty() && request.validateToken()) {
+            return companyService.deleteCompany(id);
+        } else {
+            throw new SecurityException(ERROR_MESSAGE);
+        }
+
     }
 
 }
